+++ src/main/java/com/contractor/util/Utils.java
@@ -1,0 +1,14 @@
+package com.contractor.util;
+
+import java.sql.Timestamp;
+
+public class Utils {
+
+    public static Timestamp timestampFromMillis() {
+        return Timestamp.valueOf(String.valueOf(System.currentTimeMillis()));
+    }
+
+    public static boolean compareTimestamp(Timestamp compare, Timestamp reference) {
+        return compare.before(reference) || compare.equals(reference);
+    }
+}